refactor(preprocess-helpers): migrate to vanilla JS
