+++ .changeset/forty-trains-mix.md
@@ -1,0 +1,5 @@
+---
+'@svelte-put/preprocess-helpers': major
+---
+
+Migrated to vanilla JS (tracked #203)
