+++ packages/preprocessors/helpers/.gitignore
@@ -1,0 +1,1 @@
+types/
