+++ packages/preprocessors/helpers/package.json
@@ -2,14 +2,26 @@
   "name": "@svelte-put/preprocess-helpers",
   "version": "0.0.1",
   "description": "utilities & types for building preprocessors",
-  "type": "module",
-  "main": "src/index.ts",
-  "module": "src/index.ts",
-  "types": "src/index.ts",
   "private": true,
+  "main": "src/index.js",
+  "module": "src/index.js",
+  "types": "types/index.d.ts",
+  "type": "module",
+  "exports": {
+    ".": {
+      "import": "./src/index.js",
+      "types": "./types/index.d.ts"
+    }
+  },
+  "files": [
+    "src",
+    "types"
+  ],
   "scripts": {
     "lint": "eslint --ignore-path .gitignore \"./**/*/*{ts,js,cjs}\"",
-    "format": "prettier --check --ignore-path .gitignore \"./**/*.{ts,js,cjs,json}\""
+    "format": "prettier --check --ignore-path .gitignore \"./**/*.{ts,js,cjs,json}\"",
+    "dts": "dts-buddy",
+    "prepublishOnly": "pnpm dts"
   },
   "author": "Quang Phan",
   "license": "MIT",
@@ -19,10 +31,8 @@
   "homepage": "https://github.com/vnphanquang/svelte-put/tree/main/packages/preprocessors/auto-slug",
   "devDependencies": {
     "@types/estree": "^1.0.0",
-    "@types/node": "^18.15.11",
-    "npm-run-all": "^4.1.5",
-    "svelte": "^3.58.0",
-    "typescript": "^5.0.3"
+    "@svelte-put/tsconfig": "workspace:*",
+    "dts-buddy": "^0.1.7"
   },
   "peerDependencies": {
     "svelte": "^3.55.0 || ^4.0.0"
