+++ packages/preprocessors/helpers/tsconfig.json
@@ -1,17 +1,4 @@
 {
-  "$schema": "https://json.schemastore.org/tsconfig",
-  "include": ["src/**/*.ts"],
-  "compilerOptions": {
-    "lib": ["esnext", "DOM"],
-    "isolatedModules": true,
-    "moduleResolution": "Node",
-    "esModuleInterop": true,
-    "module": "ESNext",
-    "target": "es6",
-    "sourceMap": true,
-    "declaration": true,
-    "declarationMap": true,
-    "outDir": "./dist",
-    "declarationDir": "./dist"
-  }
+  "include": ["src/**/*"],
+  "extends": "@svelte-put/tsconfig/base.package.js.json"
 }
